verification/requirements.txt
kubernetes>=28.1.0
google-cloud-compute>=1.14.0
google-cloud-container>=2.32.0
google-api-python-client>=2.108.0
pyyaml>=6.0.1
click>=8.1.7
rich>=13.7.0
requests>=2.31.0
